+++ test/Z/Data/JSON/ValueSpec.hs
@@ -13,6 +13,7 @@ import           Test.QuickCheck.Function
 import           Test.QuickCheck.Property
 import           Test.Hspec
 import           Test.Hspec.QuickCheck
+import qualified Data.Scientific as Sci
 import qualified Z.Data.JSON.Value as JSON
 import qualified Z.Data.JSON.Builder as JSONB
 
@@ -21,3 +22,9 @@ spec :: Spec
 spec = describe "JSON" $ do -- large size will generate too huge JSON document
     prop "value roundtrip" $ \ v ->
         Right v === JSON.parseValue' (B.build (JSONB.value v))
+
+    describe "floatToScientific, doubleToScientific === fromFloatDigits"  $ do
+        prop "floatToScientific == fromFloatDigits" $ \ i ->
+            JSON.floatToScientific i === Sci.fromFloatDigits i
+        prop "floatToScientific === fromFloatDigits" $ \ i ->
+            JSON.doubleToScientific i === Sci.fromFloatDigits i